Split persistent storage configuration off of the main config
