+++ src/main/java/com/bbaga/githubscheduledreminderapp/application/Config.java
@@ -5,7 +5,6 @@ import com.bbaga.githubscheduledreminderapp.domain.configuration.ConfigGraphNode
 import com.bbaga.githubscheduledreminderapp.domain.configuration.ConfigGraphUpdater;
 import com.bbaga.githubscheduledreminderapp.infrastructure.configuration.InRepoConfigParser;
 import com.bbaga.githubscheduledreminderapp.domain.configuration.RepositoryInstallationEventListener;
-import com.bbaga.githubscheduledreminderapp.infrastructure.configuration.persitance.ConfigPersistenceFactory;
 import com.bbaga.githubscheduledreminderapp.infrastructure.configuration.persitance.ConfigPersistenceInterface;
 import com.bbaga.githubscheduledreminderapp.infrastructure.github.GitHubBuilderFactory;
 import com.bbaga.githubscheduledreminderapp.infrastructure.github.webhook.EventPublisher;
@@ -38,9 +37,6 @@ import org.springframework.web.client.RestTemplate;
 
 import java.io.FileInputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 
 @Configuration
@@ -69,24 +65,6 @@ public class Config {
 
     @Value("${application.jobs.github.installationScan.interval}")
     private String installationScanIntervals;
-
-    @Value("${application.state.storage.fs.filepath}")
-    private String stateStorageFsFilePath;
-
-    @Value("${application.state.storage.type}")
-    private String stateStorageType;
-
-    @Value("${application.state.storage.gcs_bucket.name}")
-    private String stateStorageGCSBucketName;
-
-    @Value("${application.state.storage.gcs_bucket.secretFile}")
-    private String stateStorageGCSBucketSecretFile;
-
-    @Value("${application.state.storage.gcs_bucket.secret}")
-    private String stateStorageGCSBucketSecret;
-
-    @Value("${application.state.storage.gcs_bucket.filepath}")
-    private String stateStorageGCSBucketFilePath;
 
     @Bean
     public GitHub getGitHubClient(GitHubBuilderFactory gitHubBuilderFactory) throws Exception {
@@ -181,36 +159,6 @@ public class Config {
     @Qualifier("ConfigGraph")
     public ConcurrentHashMap<String, ConfigGraphNode> initializeConfigGraph(ConfigPersistenceInterface persistentConfigStorage) {
         return persistentConfigStorage.load();
-    }
-
-    @Bean
-    public ConfigPersistenceInterface getPersistentConfigStorage() {
-        ConfigPersistenceFactory factory = new ConfigPersistenceFactory();
-        HashMap<String, ?> config;
-        ConfigPersistenceFactory.PersistenceType type;
-
-        if (Objects.equals(this.stateStorageType, ConfigPersistenceFactory.PersistenceType.LOCAL_FS.label)) {
-            type = ConfigPersistenceFactory.PersistenceType.LOCAL_FS;
-            config = new HashMap<>(Map.of("filePath", stateStorageFsFilePath));
-        } else if (Objects.equals(this.stateStorageType, ConfigPersistenceFactory.PersistenceType.GCS_BUCKET.label)) {
-            type = ConfigPersistenceFactory.PersistenceType.GCS_BUCKET;
-            config = new HashMap<>(Map.of(
-                "bucketName", stateStorageGCSBucketName,
-                "credentialsJson", stateStorageGCSBucketSecret,
-                "credentialsJsonPath", stateStorageGCSBucketSecretFile,
-                "filePath", stateStorageGCSBucketFilePath
-            ));
-        } else {
-            throw new RuntimeException("Could not create state loader instance");
-        }
-
-        return factory.create(type, config);
-    }
-
-    @Bean
-    @Qualifier("application.state.storage.fs.filepath")
-    public String getStateStorageFsFilePath() {
-        return this.stateStorageFsFilePath;
     }
 
     @Bean
